[F] Use filepath from Webpack
Based on webpack-php-manifest pull request #1
https://github.com/gblair/webpack-php-manifest/pull/1

The pr features assume the webpack user will pass the relative
asset path required. This feature prevents the default asset folder
(or indeed an optional one) from being added to the files before
including. The asset folder in this plugin in this component is still
required so that October can find the webpack manifest in the build
files.